[REF] Use batch norm layers in resnet
